backend/ai_service/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pytest
httpx

